Clean routes.js
